Introduce namespace locks for all tile generation and accessing
Provides safer accessing of tiles by locking both when writing
to, but also when reading tiles from the cache.

Additionally, we limit the number of active downloads from Geonorge
to 1, to prevent running out of memory when loading several of the
large tiles to prepare composites and crop them. Each composite with
16 layers uses 1GB of memory. If we suddenly process 4/5 of them,
we get imagemagick (wand) oom errors.